Move server => cli
