+++ cli/server/proxy/contexts_test.go
