+++ cli/server/proxy/proxy.go
@@ -21,13 +21,13 @@ import (
 	"sync"
 
 	"github.com/docker/compose-cli/api/client"
+	"github.com/docker/compose-cli/cli/server/proxy/streams"
 	"github.com/docker/compose-cli/config"
 	composev1 "github.com/docker/compose-cli/protos/compose/v1"
 	containersv1 "github.com/docker/compose-cli/protos/containers/v1"
 	contextsv1 "github.com/docker/compose-cli/protos/contexts/v1"
 	streamsv1 "github.com/docker/compose-cli/protos/streams/v1"
 	volumesv1 "github.com/docker/compose-cli/protos/volumes/v1"
-	"github.com/docker/compose-cli/server/proxy/streams"
 )
 
 type clientKey struct{}
