+++ cli/server/proxy/streams/logs.go
