+++ cli/server/proxy/streams/logs_test.go
