+++ cli/server/proxy/streams/stream_test.go
