+++ cli/server/socket_windows.go
